backend/requirements.txt
Django>=5.0.6
djangorestframework-simplejwt>=5.3.1
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
django-phonenumber-field>=7.3.0
django-phonenumbers>=1.0.1
pillow>=10.3.0